+++ mappers/BusinessUnitMappers.ts
@@ -1,6 +1,7 @@
-import { BusinessUnit as CommercetoolsBusinessUnit } from '@commercetools/platform-sdk';
+import { BusinessUnit as CommercetoolsBusinessUnit, StoreKeyReference } from '@commercetools/platform-sdk';
 import { BusinessUnit } from '../types/business-unit/BusinessUnit';
 import { Store } from '../types/store/store';
+import { Associate } from '../types/associate/Associate';
 
 export class BusinessUnitMappers {
   static mapBusinessUnitToBusinessUnit(
@@ -9,90 +10,103 @@ export class BusinessUnitMappers {
     accountId: string,
     adminRoleKey: string,
   ): BusinessUnit {
-    const businessUnitWithAssociates = this.mapReferencedAssociates(businessUnit);
-
-    const businessUnitWithStores = this.mapStoreRefs(businessUnitWithAssociates, allStores);
-
-    const businessUnitWithFlags = accountId
-      ? this.addBUsinessUnitAdminFlags(businessUnitWithStores, accountId, adminRoleKey)
-      : businessUnitWithStores;
-
-    return this.trimBusinessUnit(businessUnitWithFlags);
-  }
-
-  static trimBusinessUnit(businessUnit: BusinessUnit): BusinessUnit {
     return {
       topLevelUnit: businessUnit.topLevelUnit,
       key: businessUnit.key,
-      stores: businessUnit.stores,
       name: businessUnit.name,
-      isRootAdmin: businessUnit.isRootAdmin,
-      isAdmin: businessUnit.isAdmin,
       parentUnit: businessUnit.parentUnit,
       storeMode: businessUnit.storeMode,
-      associates: businessUnit.associates?.map((associate) => ({
-        associateRoleAssignments: associate.associateRoleAssignments?.map((role) => ({
-          associateRole: { key: role.associateRole.key },
-        })),
-        customer: { id: associate.customer.id },
-      })),
+      stores: this.mapReferencedStoresToStores(businessUnit, allStores),
+      associates: this.mapReferencedAssociatesToAssociate(businessUnit),
+      isAdmin: this.isUserAdminInBusinessUnit(businessUnit, accountId, adminRoleKey),
+      isRootAdmin: this.isUserRootAdminInBusinessUnit(businessUnit, accountId, adminRoleKey),
     };
   }
 
-  static isUserAdminInBusinessUnit(businessUnit: BusinessUnit, accountId: string, adminRoleKey: string): boolean {
+  static mapBusinessUnitToBusinessUnitTreeItem(
+    businessUnit: CommercetoolsBusinessUnit,
+    allStores: Store[],
+  ): BusinessUnit {
+    return {
+      topLevelUnit: businessUnit.topLevelUnit,
+      key: businessUnit.key,
+      name: businessUnit.name,
+      parentUnit: businessUnit.parentUnit,
+      storeMode: businessUnit.storeMode,
+      stores: this.mapReferencedStoresToStores(businessUnit, allStores),
+      associates: this.mapReferencedAssociatesToAssociate(businessUnit),
+      contactEmail: businessUnit.contactEmail,
+      unitType: businessUnit.unitType,
+      custom: businessUnit.custom,
+      status: businessUnit.status,
+      addresses: businessUnit.addresses,
+    };
+  }
+
+  static trimBusinessUnit(businessUnit: BusinessUnit, accountId: string): BusinessUnit {
+    return {
+      ...businessUnit,
+      stores: businessUnit.stores.map((store) => ({ key: store.key, typeId: 'store' })),
+      associates: businessUnit.associates
+        ?.filter((associate) => associate.customer.id === accountId)
+        ?.map((associate) => ({
+          associateRoleAssignments: associate.associateRoleAssignments?.map((role) => ({
+            associateRole: { key: role.associateRole.key },
+          })),
+          customer: { id: associate.customer.id },
+        })),
+    };
+  }
+
+  static isUserAdminInBusinessUnit(
+    businessUnit: CommercetoolsBusinessUnit,
+    accountId: string,
+    adminRoleKey: string,
+  ): boolean {
     const currentUserAssociate = businessUnit.associates?.find((associate) => associate.customer.id === accountId);
     return currentUserAssociate?.associateRoleAssignments.some((role) => role.associateRole.key === adminRoleKey);
   }
 
-  static isUserRootAdminInBusinessUnit(businessUnit: BusinessUnit, accountId: string, adminRoleKey: string): boolean {
+  static isUserRootAdminInBusinessUnit(
+    businessUnit: CommercetoolsBusinessUnit,
+    accountId: string,
+    adminRoleKey: string,
+  ): boolean {
     if (this.isUserAdminInBusinessUnit(businessUnit, accountId, adminRoleKey)) {
       return !businessUnit.parentUnit;
     }
     return false;
   }
 
-  static addBUsinessUnitAdminFlags(businessUnit: BusinessUnit, accountId = '', adminRoleKey: string): BusinessUnit {
-    businessUnit.isAdmin = this.isUserAdminInBusinessUnit(businessUnit, accountId, adminRoleKey);
-    businessUnit.isRootAdmin = this.isUserRootAdminInBusinessUnit(businessUnit, accountId, adminRoleKey);
-    return businessUnit;
+  static mapReferencedAssociatesToAssociate(businessUnit: CommercetoolsBusinessUnit): Associate[] {
+    return businessUnit.associates?.map((associate) => {
+      if (associate.customer?.obj) {
+        return {
+          associateRoleAssignments: associate.associateRoleAssignments,
+          customer: {
+            id: associate.customer.id,
+            typeId: 'customer',
+            firstName: associate.customer?.obj?.firstName,
+            lastName: associate.customer?.obj?.lastName,
+            email: associate.customer?.obj?.email,
+          },
+        };
+      }
+      return associate;
+    });
   }
 
-  static mapReferencedAssociates(businessUnit: CommercetoolsBusinessUnit): BusinessUnit {
-    return {
-      ...businessUnit,
-      associates: businessUnit.associates?.map((associate) => {
-        if (associate.customer?.obj) {
-          return {
-            // @ts-ignore
-            associateRoleAssignments: associate.associateRoleAssignments,
-            customer: {
-              id: associate.customer.id,
-              typeId: 'customer',
-              firstName: associate.customer?.obj?.firstName,
-              lastName: associate.customer?.obj?.lastName,
-              email: associate.customer?.obj?.email,
-            },
-          };
-        }
-        return associate;
-      }),
-    };
-  }
-
-  static mapStoreRefs(businessUnit: BusinessUnit, allStores: Store[]): BusinessUnit {
-    return {
-      ...businessUnit,
-      stores: businessUnit.stores?.map((store) => {
-        const storeObj = allStores.find((s) => s.key === store.key);
-        return storeObj
-          ? {
-              name: storeObj.name,
-              key: storeObj.key,
-              typeId: 'store',
-              id: storeObj.id,
-            }
-          : store;
-      }),
-    };
+  static mapReferencedStoresToStores(businessUnit: CommercetoolsBusinessUnit, allStores: Store[]): StoreKeyReference[] {
+    return businessUnit.stores?.map((store) => {
+      const storeObj = allStores.find((s) => s.key === store.key);
+      return storeObj
+        ? {
+            name: storeObj.name,
+            key: storeObj.key,
+            typeId: 'store',
+            id: storeObj.id,
+          }
+        : (store as StoreKeyReference);
+    });
   }
 }